+++ projects/js-packages/ai-client/src/index.ts
@@ -1,0 +1,32 @@
+/*
+ * Core library exports
+ */
+export { default as requestJwt } from './jwt';
+export { default as SuggestionsEventSource } from './suggestions-event-source';
+export { default as askQuestion } from './ask-question';
+
+/*
+ * Hooks
+ */
+export { default as useAiSuggestions } from './hooks/use-ai-suggestions';
+export { default as useMediaRecording } from './hooks/use-media-recording';
+
+/*
+ * Components: Icons
+ */
+export * from './icons';
+
+/*
+ * Components
+ */
+export * from './components';
+
+/*
+ * Contexts
+ */
+export * from './data-flow';
+
+/*
+ * Types
+ */
+export * from './types';