+++ projects/plugins/jetpack/extensions/store/wordpress-com/reducer.ts
@@ -1,11 +1,17 @@
 /**
  * Types & Constants
  */
+import { __ } from '@wordpress/i18n';
 import {
+	ACTION_DECREASE_NEW_ASYNC_REQUEST_COUNTDOWN,
+	ACTION_ENQUEUE_ASYNC_REQUEST,
 	ACTION_INCREASE_AI_ASSISTANT_REQUESTS_COUNT,
 	ACTION_REQUEST_AI_ASSISTANT_FEATURE,
 	ACTION_SET_PLANS,
+	ACTION_SET_AI_ASSISTANT_FEATURE_REQUIRE_UPGRADE,
 	ACTION_STORE_AI_ASSISTANT_FEATURE,
+	ASYNC_REQUEST_COUNTDOWN_INIT_VALUE,
+	FREE_PLAN_REQUESTS_LIMIT,
 } from './constants';
 import type { PlanStateProps } from './types';
 
@@ -13,28 +19,34 @@ const INITIAL_STATE: PlanStateProps = {
 	plans: [],
 	features: {
 		aiAssistant: {
-			hasFeature: false,
+			hasFeature: true,
 			isOverLimit: false,
 			requestsCount: 0,
-			requestsLimit: 0,
+			requestsLimit: FREE_PLAN_REQUESTS_LIMIT,
 			requireUpgrade: false,
 			errorMessage: '',
 			errorCode: '',
 			upgradeType: 'default',
 			currentTier: {
-				value: 1,
+				slug: 'ai-assistant-tier-free',
+				value: 0,
+				limit: 20,
 			},
 			usagePeriod: {
-				currentStart: '',
+				currentStart: 'ai-assistant-tier-free',
 				nextStart: '',
 				requestsCount: 0,
-			},
-			_meta: {
-				isRequesting: false,
 			},
 			nextTier: {
 				slug: 'ai-assistant-tier-unlimited',
 				value: 1,
+				limit: 922337203685477600,
+				readableLimit: __( 'Unlimited', 'jetpack' ),
+			},
+			_meta: {
+				isRequesting: false,
+				asyncRequestCountdown: ASYNC_REQUEST_COUNTDOWN_INIT_VALUE,
+				asyncRequestTimerId: 0,
 			},
 		},
 	},
@@ -58,6 +70,8 @@ export default function reducer( state = INITIAL_STATE, action ) {
 						_meta: {
 							...state.features.aiAssistant._meta,
 							isRequesting: true,
+							asyncRequestCountdown: ASYNC_REQUEST_COUNTDOWN_INIT_VALUE, // restore the countdown
+							asyncRequestTimerId: 0, // reset the timer id
 						},
 					},
 				},
@@ -80,13 +94,104 @@ export default function reducer( state = INITIAL_STATE, action ) {
 		}
 
 		case ACTION_INCREASE_AI_ASSISTANT_REQUESTS_COUNT: {
+			// Usage Period data
+			const usagePeriod = state.features.aiAssistant.usagePeriod || { requestsCount: 0 };
+
+			// Increase requests counters
+			const requestsCount = state.features.aiAssistant.requestsCount + action.count;
+			usagePeriod.requestsCount += action.count;
+
+			// Current tier value
+			const currentTierValue = state.features.aiAssistant.currentTier?.value;
+
+			const isFreeTierPlan =
+				( typeof currentTierValue === 'undefined' && ! state.features.aiAssistant.hasFeature ) ||
+				currentTierValue === 0;
+
+			const isUnlimitedTierPlan =
+				( typeof currentTierValue === 'undefined' && state.features.aiAssistant.hasFeature ) ||
+				currentTierValue === 1;
+
+			// Request limit defined with the current tier limit by default.
+			let requestsLimit = state.features.aiAssistant.currentTier?.limit;
+
+			if ( isUnlimitedTierPlan ) {
+				requestsLimit = Infinity;
+			} else if ( isFreeTierPlan ) {
+				requestsLimit = state.features.aiAssistant.requestsLimit;
+			}
+
+			const currentCount =
+				isUnlimitedTierPlan || isFreeTierPlan // @todo: update once tier data is available
+					? requestsCount
+					: state.features.aiAssistant.usagePeriod?.requestsCount;
+
+			/**
+			 * Compute the AI Assistant Feature data optimistically,
+			 * based on the Jetpack_AI_Helper::get_ai_assistance_feature() helper.
+			 *
+			 * @see _inc/lib/class-jetpack-ai-helper.php
+			 */
+			const isOverLimit = currentCount >= requestsLimit;
+			const requireUpgrade = isOverLimit;
+
 			return {
 				...state,
 				features: {
 					...state.features,
 					aiAssistant: {
 						...state.features.aiAssistant,
-						requestsCount: state.features.aiAssistant.requestsCount + action.count,
+						isOverLimit,
+						requestsCount,
+						requireUpgrade,
+						usagePeriod: { ...usagePeriod },
+					},
+				},
+			};
+		}
+
+		case ACTION_DECREASE_NEW_ASYNC_REQUEST_COUNTDOWN: {
+			return {
+				...state,
+				features: {
+					...state.features,
+					aiAssistant: {
+						...state.features.aiAssistant,
+						_meta: {
+							...state.features.aiAssistant._meta,
+							asyncRequestCountdown: state.features.aiAssistant._meta.asyncRequestCountdown - 1,
+						},
+					},
+				},
+			};
+		}
+
+		case ACTION_ENQUEUE_ASYNC_REQUEST: {
+			return {
+				...state,
+				features: {
+					...state.features,
+					aiAssistant: {
+						...state.features.aiAssistant,
+						_meta: {
+							...state.features.aiAssistant._meta,
+							asyncRequestTimerId: action.timerId,
+						},
+					},
+				},
+			};
+		}
+
+		case ACTION_SET_AI_ASSISTANT_FEATURE_REQUIRE_UPGRADE: {
+			return {
+				...state,
+				features: {
+					...state.features,
+					aiAssistant: {
+						...state.features.aiAssistant,
+						requireUpgrade: action.requireUpgrade,
+						hasFeature: ! action.requireUpgrade, // If we require an upgrade, we don't have the feature.
+						isOverLimit: true, // If we require an upgrade, we are over the limit.
 					},
 				},
 			};