+++ projects/plugins/jetpack/extensions/store/wordpress-com/types.ts
@@ -8,15 +8,15 @@ export type UpgradeTypeProp = 'vip' | 'default';
 
 export type TierUnlimitedProps = {
 	slug: 'ai-assistant-tier-unlimited';
-	limit: 9223372036854776000n;
+	limit: number;
 	value: 1;
-	readableLimit: 'Unlimited';
+	readableLimit: string;
 };
 
 export type TierFreeProps = {
 	slug: 'ai-assistant-tier-free';
 	limit: 20;
-	value: 20;
+	value: 0;
 };
 
 export type Tier100Props = {
@@ -37,12 +37,27 @@ export type Tier500Props = {
 	value: 500;
 };
 
+export type TierProp = {
+	slug: TierSlugProp;
+	limit: TierLimitProp;
+	value: TierValueProp;
+	readableLimit?: string;
+};
+
+export type TierLimitProp =
+	| TierUnlimitedProps[ 'limit' ]
+	| TierFreeProps[ 'limit' ]
+	| Tier100Props[ 'limit' ]
+	| Tier200Props[ 'limit' ]
+	| Tier500Props[ 'limit' ];
+
 export type TierSlugProp =
 	| TierUnlimitedProps[ 'slug' ]
 	| TierFreeProps[ 'slug' ]
 	| Tier100Props[ 'slug' ]
 	| Tier200Props[ 'slug' ]
 	| Tier500Props[ 'slug' ];
+
 export type TierValueProp =
 	| TierUnlimitedProps[ 'value' ]
 	| TierFreeProps[ 'value' ]
@@ -56,29 +71,24 @@ export type AiFeatureProps = {
 	requestsCount: number;
 	requestsLimit: number;
 	requireUpgrade: boolean;
-	errorMessage: string;
-	errorCode: string;
+	errorMessage?: string;
+	errorCode?: string;
 	upgradeType: UpgradeTypeProp;
-	currentTier: {
-		value: TierValueProp;
-	};
-	usagePeriod: {
+	currentTier?: TierProp;
+	usagePeriod?: {
 		currentStart: string;
 		nextStart: string;
 		requestsCount: number;
 	};
-	nextTier: {
-		slug: TierSlugProp;
-		value: TierValueProp;
-		limit?: number;
-		redeableLimit?: string;
-	};
+	nextTier?: TierProp | null;
 };
 
 // Type used in the `wordpress-com/plans` store.
 export type AiFeatureStateProps = AiFeatureProps & {
 	_meta?: {
 		isRequesting: boolean;
+		asyncRequestCountdown: number;
+		asyncRequestTimerId: number;
 	};
 };