+++ projects/plugins/jetpack/extensions/types.ts
@@ -1,12 +1,13 @@
 /**
  * Types for the AI Assistant feature.
  */
-import type { TierSlugProp, TierValueProp, UpgradeTypeProp } from './store/wordpress-com/types';
+import type { TierProp, UpgradeTypeProp } from './store/wordpress-com/types';
 
 /*
  * `sites/$site/ai-assistant-feature` endpoint response body props
  */
 export type SiteAIAssistantFeatureEndpointResponseProps = {
+	'is-enabled': boolean;
 	'has-feature': boolean;
 	'is-over-limit': boolean;
 	'requests-count': number;
@@ -21,18 +22,7 @@ export type SiteAIAssistantFeatureEndpointResponseProps = {
 	'error-code'?: string;
 	'is-playground-visible'?: boolean;
 	'upgrade-type': UpgradeTypeProp;
-	'current-tier': {
-		value: TierValueProp;
-	};
-	'tier-plans': Array< {
-		slug: TierSlugProp;
-		limit: number;
-		value: TierValueProp;
-	} >;
-	'next-tier'?: {
-		slug: TierSlugProp;
-		value: TierValueProp;
-		limit: number;
-		'redeable-limit': string;
-	};
+	'current-tier': TierProp;
+	'tier-plans': Array< TierProp >;
+	'next-tier'?: TierProp | null;
 };