+++ projects/plugins/jetpack/tests/php/modules/masterbar/test-class-jetpack-admin-menu.php
@@ -99,7 +99,6 @@ class Test_Jetpack_Admin_Menu extends WP_UnitTestCase {
 
 		static::$admin_menu->add_jetpack_menu();
 		$this->assertSame( 'https://wordpress.com/scan/' . static::$domain, $submenu['jetpack'][2][2] );
-		$this->assertSame( 'https://wordpress.com/subscribers/' . static::$domain, $submenu['jetpack'][3][2] );
 	}
 
 	/**